#5424: remove expm1 op from metal_ckernel_sfpu
