+++ tt_metal/hw/ckernels/wormhole_b0/metal/llk_api/llk_math_unary_sfpu_api.h
@@ -235,17 +235,6 @@ inline void llk_math_eltwise_unary_sfpu_heaviside_init() {
     llk_math_eltwise_unary_sfpu_init<SfpuType::heaviside, APPROXIMATE>();
 }
 
-//EXPM1
-template <bool APPROXIMATE, DstSync dst_sync = DstSync::SyncFull>
-inline void llk_math_eltwise_unary_sfpu_expm1(uint dst_index) {
-    llk_math_eltwise_unary_sfpu<SfpuType::expm1, APPROXIMATE, dst_sync>(dst_index);
-}
-
-template <bool APPROXIMATE>
-inline void llk_math_eltwise_unary_sfpu_expm1_init() {
-    llk_math_eltwise_unary_sfpu_init<SfpuType::expm1, APPROXIMATE>();
-}
-
 //Asin
 template <bool APPROXIMATE, DstSync dst_sync = DstSync::SyncFull>
 inline void llk_math_eltwise_unary_sfpu_asin(uint dst_index) {
