+++ tt_metal/hw/ckernels/wormhole_b0/metal/llk_api/llk_sfpu/llk_math_eltwise_unary_sfpu_common_includes.h
@@ -88,8 +88,6 @@ inline void llk_math_calculate_sfpu(
         calculate_exp2<APPROXIMATION_MODE, ITERATIONS>();
     } else if constexpr (operation == SfpuType::heaviside) {
         calculate_heaviside<APPROXIMATION_MODE, ITERATIONS>(param0);
-    } else if constexpr (operation == SfpuType::expm1) {
-        calculate_expm1<APPROXIMATION_MODE, ITERATIONS>();
     } else if constexpr (operation == SfpuType::asin) {
         calculate_asin<APPROXIMATION_MODE, ITERATIONS>();
     } else if constexpr (operation == SfpuType::acos) {
